requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
datasets==4.4.2
dill==0.4.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
lingua-language-detector==1.3.4
MarkupSafe==3.0.3
multidict==6.7.0
multiprocess==0.70.18
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pip==25.3
propcache==0.4.1
pyarrow==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2023.12.25
requests==2.32.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
xxhash==3.6.0
yarl==1.22.0

chromadb